requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pymongo==4.6.0

# Data Analysis
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.1
kaleido==0.2.1  # For static image export in Plotly

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2

# Optional: For advanced features
# scikit-learn==1.3.2  # Para ML futuro
# streamlit==1.29.0    # Para dashboard interactivo
